# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: StockPilot - Inventory & POS SaaS
  description: '

    ## StockPilot API - Inventory & POS SaaS


    Multi-tenant inventory management, point-of-sale, finance tracking, and AI insights

    for small businesses.


    ### Authentication

    All API endpoints require a Bearer token in the Authorization header:

    `Authorization: Bearer <access_token>`


    Include the business ID in the X-Business-Id header:

    `X-Business-Id: <business_id>`


    ### Rate Limiting

    - Default: 100 requests per minute per user

    - Auth endpoints: 5 requests per minute per IP


    ### Support

    For issues and feature requests, contact support.

    '
  version: 1.0.0
paths:
...
  /api/v1/suppliers/{sid}/purchases:
    get:
      tags:
      - parties
      - suppliers
      summary: Sup History
      operationId: sup_history_api_v1_suppliers__sid__purchases_get
      security:
      - HTTPBearer: []
      parameters:
      - name: sid
        in: path
        required: true
        schema:
          type: integer
          title: Sid
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/purchases:
    get:
      tags:
      - purchases
      summary: List All
      operationId: list_all_api_v1_purchases_get
      security:
      - HTTPBearer: []
      parameters:
      - name: supplier_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Supplier Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PurchaseOut'
                title: Response List All Api V1 Purchases Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - purchases
      summary: Create
      description: 'FR-10: multi-item purchase. Totals auto-calculated, stock up on confirm.'
      operationId: create_api_v1_purchases_post
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/purchases/{pid}/pay:
    post:
      tags:
      - purchases
      summary: Pay
      description: 'FR-10.8: record payment against a purchase, update supplier balance.'
      operationId: pay_api_v1_purchases__pid__pay_post
      security:
      - HTTPBearer: []
      parameters:
      - name: pid
        in: path
        required: true
        schema:
          type: integer
          title: Pid
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchasePayRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/purchases/{pid}/cancel:
    post:
      tags:
      - purchases
      summary: Cancel
      description: 'Cancel a purchase: reverse added stock, restore supplier balance.'
      operationId: cancel_api_v1_purchases__pid__cancel_post
      security:
      - HTTPBearer: []
      parameters:
      - name: pid
        in: path
        required: true
        schema:
          type: integer
          title: Pid
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/purchases/{pid}:
    get:
      tags:
      - purchases
      summary: Get One
      operationId: get_one_api_v1_purchases__pid__get
      security:
      - HTTPBearer: []
      parameters:
      - name: pid
        in: path
        required: true
        schema:
          type: integer
          title: Pid
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/reports/purchases:
    get:
      tags:
      - reports
      summary: Purchase Report
      description: 'FR-23.3/23.7/23.8: purchase report.'
      operationId: purchase_report_api_v1_reports_purchases_get
      security:
      - HTTPBearer: []
      parameters:
      - name: preset
        in: query
        required: false
        schema:
          type: string
          default: month
          title: Preset
      - name: format
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Format
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100000
          minimum: 1
          default: 2000
          title: Limit
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PurchaseCreate:
      properties:
        supplier_id:
          type: integer
          title: Supplier Id
        items:
          items:
            $ref: '#/components/schemas/PurchaseItemIn'
          type: array
          minItems: 1
          title: Items
        discount_amount:
          type: number
          title: Discount Amount
          default: 0
        tax_amount:
          type: number
          title: Tax Amount
          default: 0
        paid_amount:
          type: number
          title: Paid Amount
          default: 0
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - supplier_id
      - items
      title: PurchaseCreate
    PurchaseItemIn:
      properties:
        product_id:
          type: integer
          title: Product Id
        quantity:
          type: integer
          exclusiveMinimum: 0.0
          title: Quantity
        unit_cost:
          type: number
          minimum: 0.0
          title: Unit Cost
      type: object
      required:
      - product_id
      - quantity
      - unit_cost
      title: PurchaseItemIn
    PurchaseItemOut:
      properties:
        id:
          type: integer
          title: Id
        product_id:
          type: integer
          title: Product Id
        product_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Name
        quantity:
          type: integer
          title: Quantity
        unit_cost:
          type: number
          title: Unit Cost
        line_total:
          type: number
          title: Line Total
      type: object
      required:
      - id
      - product_id
      - quantity
      - unit_cost
      - line_total
      title: PurchaseItemOut
    PurchaseOut:
      properties:
        id:
          type: integer
          title: Id
        supplier_id:
          type: integer
          title: Supplier Id
        supplier_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Supplier Name
        purchase_date:
          type: string
          format: date-time
          title: Purchase Date
        subtotal:
          type: number
          title: Subtotal
        discount_amount:
          type: number
          title: Discount Amount
        tax_amount:
          type: number
          title: Tax Amount
        total_amount:
          type: number
          title: Total Amount
        paid_amount:
          type: number
          title: Paid Amount
        payment_status:
          type: string
          title: Payment Status
        status:
          type: string
          title: Status
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        items:
          items:
            $ref: '#/components/schemas/PurchaseItemOut'
          type: array
          title: Items
          default: []
      type: object
      required:
      - id
      - supplier_id
      - purchase_date
      - subtotal
      - discount_amount
      - tax_amount
      - total_amount
      - paid_amount
      - payment_status
      - status
      title: PurchaseOut
    PurchasePayRequest:
      properties:
        amount:
          type: number
          exclusiveMinimum: 0.0
          title: Amount
      type: object
      required:
      - amount
      title: PurchasePayRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer